requirements/common.txt
Django==1.5.4
South==0.8.2
Unidecode==0.04.13
facebook-sdk==0.4.0
facepy==0.8.4
requests==1.2.3
wsgiref==0.1.2
